refactored modals so I can reuse the logic and reduce code
